QRCode: improve isPure detection
Make the 'pure' detector capable of handling symbols with a non-integer
module width. This can happen due to aliasing if the rendered symbol
width in pixels is not an integer multiple of the symbol width in modules.

This finds 4 more samples from the blackbox test set and also the example
from the KDE project mentioned in
https://github.com/nu-book/zxing-cpp/issues/128#issuecomment-730479329.

+++ core/src/BitMatrix.h
@@ -303,7 +303,7 @@ BitMatrix Inflate(BitMatrix&& input, int width, int height, int quietZone);
  * @param subSampling typically the module size
  * @return deflated input
  */
-BitMatrix Deflate(const BitMatrix& matrix, int width, int height, int top, int left, int subSampling);
+BitMatrix Deflate(const BitMatrix& matrix, int width, int height, float top, float left, float subSampling);
 
 template<typename T>
 BitMatrix ToBitMatrix(const Matrix<T>& in, T trueValue = {true})
